I format songs to be easier to see.
