+++ components/HomePage/HomePage.tsx
@@ -1,25 +1,9 @@
 import { ConnectButton } from '@rainbow-me/rainbowkit'
-import { useEffect, useState } from 'react'
-import { getNfts } from '../../lib/getNfts'
-import { useAccount, useNetwork } from 'wagmi'
-import Song from '@components/Song'
+import { useAccount } from 'wagmi'
+import SongList from '../SongList'
 
 const HomePage = () => {
-  const { activeChain } = useNetwork()
   const { data: account } = useAccount()
-  const [songs, setSongs] = useState([] as any)
-
-  useEffect(() => {
-    const init = async () => {
-      console.log('SWEETS GETTING MUSIC')
-      const response = await getNfts(activeChain?.id?.toString(), account.address)
-      console.log('SWEETS RESPONSE', response)
-      setSongs(response)
-    }
-
-    if (!activeChain || !account) return
-    init()
-  }, [activeChain, account])
 
   return (
     <div className="flex flex-col justify-center items-center h-[100vh] gap-11">
@@ -29,7 +13,7 @@ const HomePage = () => {
         across the Ethereum ecosystem.
       </div>
       <ConnectButton label="sign in" />
-      {account && songs && songs.map((song) => <Song key={song?.title} song={song} />)}
+      {account && <SongList />}
     </div>
   )
 }
