+++ components/SongList/index.ts
@@ -1,0 +1,3 @@
+import SongList from "./SongList";
+
+export default SongList
