+++ lib/getIpfsLink.ts
@@ -1,0 +1,6 @@
+const getIpfsLink = (hash: string) =>
+  hash?.indexOf?.("ipfs://") > -1
+    ? hash.replace("ipfs://", "https://nftstorage.link/ipfs/")
+    : hash;
+
+export default getIpfsLink;
